requirements.txt
click==7.1.2
Pillow==7.1.2
pycodestyle==2.6.0
